Replace instances of `python3` with `python` Helpful for this issue: https://stackoverflow.com/questions/64801225/python-or-python3-what-is-the-difference
Where only python 3.x is installed, `python` is the command to use.
Otherwise, this error arises: `Python was not found; run without arguments to install from the Microsoft Store, or disable this shortcut from Settings > Manage App Execution Aliases.`

Where both python 2.x and python 3.x are installed, `python3` is the command to use.
This avoids the above error.